requirements.txt
aiohttp>=3.8.0
python-dotenv>=1.0.0
byteplus-python-sdk-v2
google-genai
Pillow>=10.0.0